Add : 주문 후기 리스트 조회

+++ android/app/src/main/java/com/ssafy/icepop/data/model/dto/IceCreamOrderReview.kt
@@ -6,5 +6,5 @@ data class IceCreamOrderReview(
     val memberEmail: String,
     val name: String,
     val orderId: Int,
-    val rate: Int
+    val rate: Double
 )
